% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwebq.R
\name{.webq_request}
\alias{.webq_request}
\title{.webq_request}
\usage{
.webq_request(endpoint, ...)
}
\arguments{
\item{endpoint}{passed to GET as \code{url}}

\item{...}{additional parameters passed to GET}
}
\value{
HTML Response as a character string
}
\description{
.webq_request
}
