% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwebq.R
\name{webq_config}
\alias{webq_config}
\title{webq_config}
\usage{
webq_config(netid, key)
}
\arguments{
\item{netid}{Your UW Netid}

\item{key}{Your Catalyst API Key}
}
\description{
You must run this function before you can use the other features of this package.
If you don't have a Catalyst API key, get one here: https://catalyst.uw.edu/rest_user.
}
\examples{
webq_config('susieq', 'myapikey000111222333')

}
