% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwebq.R
\name{webq_responses}
\alias{webq_responses}
\title{webq_responses}
\usage{
webq_responses(survey_id, participant_ids = NULL)
}
\arguments{
\item{survey_id}{see \code{webq_surveys} to identify \code{survey_id}}

\item{participant_ids}{(optional) see \code{webq_participants} to identify \code{participant_id}}
}
\value{

}
\description{
Returns a tibble of responses the responses for a given survey and,
if provided, given participants. The tibble includes the following rows:
\itemize{
\item question_id
\item question_type
\item question_content: the prompt on the Catalyst survey
\item text_values: the participant's responses if text type
\item numeric_values: the participant's responses if numeric type
\item participant_id
}
}
\examples{
\dontrun{webq_responses('236008', participant_ids = c('20700916', '12996001'))}

}
