% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwebq.R
\name{webq_surveys}
\alias{webq_surveys}
\title{webq_surveys}
\usage{
webq_surveys()
}
\value{
a tibble
}
\description{
Returns a tibble of WebQ surveys that the user is authorized to view.
The tibble has the following rows:
\itemize{
\item name: the survey name
\item survey_id: the id that can be used as an argument to retrieve results with \code{webq_survey_results}
\item response_count: the current number of responses
}
}
\examples{
webq_surveys()
}
